Convert LWRPs to HWRPs to preserve names
Per https://github.com/lusis/chef-kibana/issues/61, preserving the names of the LWRPs while potentially having to change the cookbook name to get it posted to Supermarket. This PR converts the LWRPs in the cookbook into proper HWRPs using @coderanger's [poise helper library](https://github.com/poise/poise) for writing heavy weight resources and providers.

I haven't really changed anything except reorganized it. This should allow us to post the cookbook to Supermarket under another name while preserving the resource calls in downstream recipes.
